+++ lib/reputation_system/evaluation_methods.rb
@@ -15,7 +15,7 @@
 ##
 
 module ReputationSystem
-  module Evaluation
+  module EvaluationMethods
     module ClassMethods
       def evaluated_by(reputation_name, source, *args)
         scope = args.first
@@ -37,9 +37,9 @@ module ReputationSystem
       scope = args.first
       srn = ReputationSystem::Network.get_scoped_reputation_name(self.class.name, reputation_name, scope)
       process = ReputationSystem::Network.get_reputation_def(self.class.name, srn)[:aggregated_by]
-      evaluation = RSEvaluation.create_evaluation(srn, value, source, self)
-      rep = RSReputation.find_or_create_reputation(srn, self, process)
-      RSReputation.update_reputation_value_with_new_source(rep, evaluation, 1, process)
+      evaluation = ReputationSystem::Evaluation.create_evaluation(srn, value, source, self)
+      rep = ReputationSystem::Reputation.find_or_create_reputation(srn, self, process)
+      ReputationSystem::Reputation.update_reputation_value_with_new_source(rep, evaluation, 1, process)
     end
 
     def update_evaluation(reputation_name, value, source, *args)
@@ -48,13 +48,13 @@ module ReputationSystem
       evaluation.value = value
       evaluation.save!
       process = ReputationSystem::Network.get_reputation_def(self.class.name, srn)[:aggregated_by]
-      rep = RSReputation.find_by_reputation_name_and_target(srn, self)
-      RSReputation.update_reputation_value_with_updated_source(rep, evaluation, oldValue, 1, process)
+      rep = ReputationSystem::Reputation.find_by_reputation_name_and_target(srn, self)
+      ReputationSystem::Reputation.update_reputation_value_with_updated_source(rep, evaluation, oldValue, 1, process)
     end
 
     def add_or_update_evaluation(reputation_name, value, source, *args)
       srn, evaluation = find_srn_and_evaluation(reputation_name, source, args.first)
-      if RSEvaluation.exists? :reputation_name => srn, :source_id => source.id, :source_type => source.class.name, :target_id => self.id, :target_type => self.class.name
+      if ReputationSystem::Evaluation.exists? :reputation_name => srn, :source_id => source.id, :source_type => source.class.name, :target_id => self.id, :target_type => self.class.name
         self.update_evaluation(reputation_name, value, source, *args)
       else
         self.add_evaluation(reputation_name, value, source, *args)
@@ -82,7 +82,7 @@ module ReputationSystem
     protected
       def find_srn_and_evaluation(reputation_name, source, scope)
         srn = ReputationSystem::Network.get_scoped_reputation_name(self.class.name, reputation_name, scope)
-        evaluation = RSEvaluation.find_by_reputation_name_and_source_and_target(srn, source, self)
+        evaluation = ReputationSystem::Evaluation.find_by_reputation_name_and_source_and_target(srn, source, self)
         return srn, evaluation
       end
 
@@ -93,7 +93,7 @@ module ReputationSystem
       end
 
       def find_evaluation!(reputation_name, srn, source)
-        evaluation = RSEvaluation.find_by_reputation_name_and_source_and_target(srn, source, self)
+        evaluation = ReputationSystem::Evaluation.find_by_reputation_name_and_source_and_target(srn, source, self)
         raise ArgumentError, "Given instance of #{source.class.name} has not evaluated #{reputation_name} of the instance of #{self.class.name} yet." unless evaluation
         evaluation
       end
@@ -102,15 +102,15 @@ module ReputationSystem
         process = ReputationSystem::Network.get_reputation_def(self.class.name, srn)[:aggregated_by]
         oldValue = evaluation.value
         evaluation.value = process == :product ? 1 : 0
-        rep = RSReputation.find_by_reputation_name_and_target(srn, self)
-        RSReputation.update_reputation_value_with_updated_source(rep, evaluation, oldValue, 1, process)
+        rep = ReputationSystem::Reputation.find_by_reputation_name_and_target(srn, self)
+        ReputationSystem::Reputation.update_reputation_value_with_updated_source(rep, evaluation, oldValue, 1, process)
         evaluation.destroy
       end
 
       def change_evaluation_value_by(reputation_name, value, source, *args)
         scope = args.first
         srn = ReputationSystem::Network.get_scoped_reputation_name(self.class.name, reputation_name, scope)
-        evaluation = RSEvaluation.find_by_reputation_name_and_source_and_target(srn, source, self)
+        evaluation = ReputationSystem::Evaluation.find_by_reputation_name_and_source_and_target(srn, source, self)
         if evaluation.nil?
           self.add_evaluation(reputation_name, value, source, scope)
         else